requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langchain>=0.1.0
langgraph>=0.0.30
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
openai>=1.6.0
anthropic>=0.8.0
gradio>=6.5.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
python-multipart>=0.0.6
langgraph-checkpoint>=0.0.1

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# RAG and Vector Database
chromadb>=0.4.0
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
